book suport

+++ Neoniks/ViewControllers/MagicViewController/MagicWorldViewController.h
@@ -9,7 +9,7 @@
 #import <UIKit/UIKit.h>
 @protocol MagicWorldDelegate <NSObject>
 -(void)close;
--(void)next:(int)pageToShow isPrev:(BOOL)prev;
+-(void)next:(NSInteger)pageToShow isPrev:(BOOL)prev;
 
 @end
